requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosmtplib==3.0.1
black==24.3.0
certifi==2024.2.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
PyPDF2==3.0.1
python-decouple==3.8
